printing numbers

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -17,7 +17,7 @@ int main(void)
 
 	for (n = 0; n < 10; ++n)
 	{
-		putchar(n);
+		putchar((int)n);
 	}
 
 	putchar('\n');
